tests/contracts/requirements-contracts.txt
# Contract Testing Dependencies
# ================================
# For Consumer-Driven Contract testing with Pact

# Pact for Python - Consumer-Driven Contracts
pact-python>=1.7.0

# Testing framework
pytest>=8.0.0
pytest-asyncio>=0.21.0

# HTTP client for contract verification
requests>=2.31.0
httpx>=0.25.0

# JSON schema validation
jsonschema>=4.19.0

# Mock server for provider testing (optional)
pactman>=0.8.0